- Refactor to be a bit easier to add more APIs to
